test: move wisp and ebollo utilities into their suite folders
