+++ src/integration/cup/utility/security.py
@@ -1,3 +1,3 @@
-from src.integration.utility.wisp.utils import obfuscate_secrets
+from src.integration.wisp.utility.utils import obfuscate_secrets
 
 __all__ = ["obfuscate_secrets"]
