+++ src/integration/wisp/steps/healthcheck.py
@@ -3,8 +3,8 @@ import logging
 from behave import given
 
 from src.conf.configuration import secrets
-from src.integration.utility.wisp import constants
-from src.integration.utility.wisp import utils
+from src.integration.wisp.utility import constants
+from src.integration.wisp.utility import utils
 
 
 @given('i sistemi sono operativi')
